refactor(tests): put tests in a test folder and all lib function in lib.rs
