+++ tests/Cargo.toml
@@ -22,4 +22,6 @@ harness = false
 
 [dev-dependencies]
 assert_cmd = "2.0.12"
+assert_fs = "1.0.13"
+predicates = "3.0.3"
 wasmparser = "0.110.0"
